The HTML::Field set of modules creates objects that represent HTML form fields which try to make it easier to interact with CGI objects, databases, and HTML::Template objects.
The objective of an HTML::Field object is to know how to write its own HTML,
how to get its value out of a CGI object or from a hash,
how to add their value to a hash suitable for passing into a HTML::Template
or into a SQL::Abstract object, for example, and thus re-use some of the code
which is typically repeated several times in a CGI script.

This bundle includes also HTML::FieldForm, which is a very simple module to
manage sets of HTML::Field objects.

WWW: http://search.cpan.org/dist/HTML-Field/

PR:		ports/140982
Submitted by:	Sahil Tandon <sahil at tandon.net>