Add port entry for the base system LPR sources.
When world is installed w/ WITHOUT_LPR, they will not be present. Like
the compat packages this allows it to be added back in; this also supports
novel uses like using pfSense to build an embedded print server.

FYI: For raw queue use, bsd lpd's footprint is a few 100k on disk; cups
is ~12MB.

Reviewed by:	dougb